level1_p02 finished.

+++ level1/p02_is_prime/main.c
@@ -4,7 +4,7 @@ int main(void){
     int num = 0;int i = 2;
     printf("Please enter a integer:");
     scanf_s("%d", &num);
-    int dividing_number =sqrt(num);
+    int dividing_number =(int)sqrt(num);
     for(;i<dividing_number+1;i++) {
         if(num%i==0) {
             printf("The number is not prime.");
